# Improve _make_vertice_cells performance