@akaszynski 2020-05-15
> Wow, I had no idea this was a bottleneck when I implemented it... is the issue with the `ascontiguousarray` call?

It's mostly that because the shape of cells/faces is no 2D, `pyvista` has no idea that it's actually a valid `vtk` cell array and it instead tries to loop through it.